port test to integrationTest
