+++ src/test/java/org/opensearch/security/dlic/rest/api/TenantInfoActionTest.java
@@ -19,19 +19,8 @@ import org.junit.Test;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.security.support.ConfigConstants;
 import org.opensearch.security.test.helper.rest.RestHelper;
-import org.opensearch.test.framework.AsyncActions;
 
 import static org.opensearch.security.OpenSearchSecurityPlugin.PLUGINS_PREFIX;
-
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.anyOf;
-import static org.hamcrest.MatcherAssert.assertThat;
 
 public class TenantInfoActionTest extends AbstractRestApiUnitTest {
     private String payload = "{\"hosts\":[],\"users\":[\"sarek\"],"
@@ -95,33 +84,5 @@ public class TenantInfoActionTest extends AbstractRestApiUnitTest {
         rh.sendAdminCertificate = false;
         response = rh.executeGetRequest(ENDPOINT);
         Assert.assertEquals(HttpStatus.SC_OK, response.getStatusCode());
-    }
-
-    @Test
-    public void testParallelPutRequests() throws Exception {
-        setup();
-
-        rh.keystore = "restapi/kirk-keystore.jks";
-        rh.sendAdminCertificate = true;
-        final String TENANTS_ENDPOINT = BASE_ENDPOINT + "/api/tenants/tenant1";
-        final String TENANTS_BODY = "{\"description\":\"create new tenant\"}";
-
-        final CountDownLatch countDownLatch = new CountDownLatch(1);
-        final List<CompletableFuture<RestHelper.HttpResponse>> conflictingRequests = AsyncActions.generate(() -> {
-            return rh.executePutRequest(TENANTS_ENDPOINT, TENANTS_BODY);
-        }, 2, 4);
-
-        // Make sure all requests start at the same time
-        countDownLatch.countDown();
-
-        AtomicInteger numCreatedResponses = new AtomicInteger();
-        AsyncActions.getAll(conflictingRequests, 1, TimeUnit.SECONDS).forEach((response) -> {
-            assertThat(response.getStatusCode(), anyOf(equalTo(HttpStatus.SC_CREATED), equalTo(HttpStatus.SC_CONFLICT)));
-            if (response.getStatusCode() == HttpStatus.SC_CREATED) numCreatedResponses.getAndIncrement();
-        });
-        assertThat(numCreatedResponses.get(), equalTo(1)); // should only be one 201
-
-        RestHelper.HttpResponse getResponse = rh.executeGetRequest(TENANTS_ENDPOINT); // make sure the one 201 works
-        assertThat(getResponse.findValueInJson("tenant1" + ".description"), equalTo("create new tenant"));
     }
 }
